refactor: move modulo GetCurrentQuadrimester para acf/shared
